Перенес журнал точек доставки, объединил в папку с новым журналом, вынес ноду в отдельный файл
