+++ Vodovoz/JournalFilters/UndeliveredOrdersFilter.cs
@@ -18,6 +18,7 @@ using Vodovoz.Journals.JournalViewModels.Organization;
 using Vodovoz.JournalViewModels;
 using Vodovoz.TempAdapters;
 using Vodovoz.ViewModels.Journals.JournalFactories;
+using Vodovoz.ViewModels.Journals.JournalViewModels.Client;
 
 namespace Vodovoz.JournalFilters
 {
