+++ Vodovoz/TempAdapters/DeliveryPointJournalFactory.cs
@@ -14,6 +14,7 @@ using Vodovoz.Filters.ViewModels;
 using Vodovoz.JournalViewModels;
 using Vodovoz.Parameters;
 using Vodovoz.ViewModels.Journals.JournalViewModels;
+using Vodovoz.ViewModels.Journals.JournalViewModels.Client;
 using Vodovoz.ViewModels.TempAdapters;
 
 namespace Vodovoz.TempAdapters
